requirements.txt
alembic
flask[async]
ipython
ollama
psycopg2-binary
python-dotenv
requests
sqlmodel
